fix: Quick fix `LineSurface` tolerance inconsistency in `globalToLocal` (#2239)
This is meant as a quick fix for the problem I am experiencing in https://github.com/acts-project/acts/pull/2086. It seems like it also appeared somewhere else already and maybe in one of the recent issues.

The problem seems to be that the tolerance is not consistent between `intersection` and `globalToLocal`. Since these are used in sequence with the same tolerance `globalToLocal` will fail as it reports a higher distance than the `intersection`.

A proper fix would be to align the math of the intersection and conversion so the tolerance has the same meaning in both cases.

should also fix https://github.com/acts-project/acts/issues/2226